Upload Linux GTK embedding artifacts
The in-progress GTK embedding will replace the GLFW embedding on Linux
once complete. This uploads it, and its headers, for all modes. Doing
so now, even though the embedding isn't complete yet, facilitates
development of the tooling and runner support, and the structure of
the artifacts as captured in the recipe isn't expected to change.

Bug: https://github.com/flutter/flutter/issues/54860
Change-Id: Ic013b1a7cb3a13a3d04a77721bd499050e4ea323
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/2261
Reviewed-by: Godofredo Contreras <godofredoc@google.com>