Fix the logic to kill processes on windows.
For windows we need to kill all the processes started by the task to
ensure the task can cleanup all the created temporary folders. The
existent logic was killing the processes but still raising and infra
failure if the task failed to kill one of the commands.

Change-Id: I9c7421b7a659fc27d753fa05b5bb907cfe083dd5
Bug: https://github.com/flutter/flutter/issues/63699
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/5803
Commit-Queue: Godofredo Contreras <godofredoc@google.com>
Reviewed-by: Keyong Han <keyonghan@google.com>