Refactor android module to remove the need for a context.
Android module was using a context even though it was not doing anything
in the cleanup. Removing the context allows to pass android as a regular
dependency in a dynamic way.

This also uses android licenses from cipd instead of creating the files
at runtime to remove non necessary interations of recipes with configs.

Change-Id: I972832e870967e6944f5558a130f1b17177d3392
Bug: https://github.com/flutter/flutter/issues/65187
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/6280
Commit-Queue: Godofredo Contreras <godofredoc@google.com>
Reviewed-by: Tong Wu <wutong@google.com>