Notify on completion
This is a resubmit of the previously reverted CL, which was crashing because
the prod builder user account was not given necessary permissions for
pubsub. This change has added error handling to avoid other failures to bring
down the job as well:

Now sending a pubsub message on recipe completion which will enable running
tests as soon as builds are complete. Added bucket as well to the message to
differentiate between prod and try jobs.

Added a try catch for StepFailure to ensure that failures in gcloud pubsub do
not bring down the overall job.

Tested locally:

  led get-builder 'luci.flutter.try:Mac Host Engine' | led edit-recipe-bundle | led launch
  led get-builder 'luci.flutter.prod:Linux Fuchsia' | led edit-recipe-bundle | led launch

Verified that failing gcloud step does not fail the whole job:

  http://shortn/_HG8zHwgztC

and verified that expected messages appeared in the topic:

  {"buildername" : "Linux Fuchsia", "bucket" : "prod", "githash" : "4985b584a430ea3fda9433dccafa5047531e0974"}

This change will be superceded by user defined pubsub topic (crbug.com/843720).

Bug: https://github.com/flutter/flutter/issues/69621
Change-Id: I424cd8963372c8351940763c8813c9439777b3ba
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/9201
Commit-Queue: Yusuf Mohsinally <mohsinally@google.com>
Reviewed-by: Keyong Han <keyonghan@google.com>