Add a cipd dependency for curl.
The curl binary in the debian linux images is very old and it doesn't
include several fixes. We will be providing the curl binary over cipd to
ensure we have a recent version of it. This is very important because
most of the flutter tooling depends on curl to download dependencies.

Change-Id: Id5ef97e492376726bf5fdb4a26a1f804d7dd42e0
Bug: https://github.com/flutter/flutter/issues/68242
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/10780
Reviewed-by: Casey Hillers <chillers@google.com>
Commit-Queue: Godofredo Contreras <godofredoc@google.com>