Web tests require both the source and artifacts for local-engine.
This tests were passing in the engine prod because most of the caches
were already populated with the correct source code version in the
expected location. In try builders there is more rotation of caches and
tests were failing if the engine build cache was not already populated.

Change-Id: I84f7afe1c3d0b3f57a5e05fb7967780f313a0dd9
Bug: https://github.com/flutter/flutter/issues/78804
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/11420
Commit-Queue: Godofredo Contreras <godofredoc@google.com>
Reviewed-by: Yegor Jbanov <yjbanov@google.com>