Timeout devicelab tasks on 10 mins.
The longest task(microbenchmarks) takes a total of 6mins to execute.
This will ensure that we fail fast and we don't waste 45 mins waiting
for the task timeout to trigger. More importantly test timeouts are
actual errors in the test and this will prevent counting them as infra
failures.

Change-Id: I9f6eeca520c158fb766553437b1f95dfeb63d080
Bug: https://github.com/flutter/flutter/issues/79993
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/12160
Reviewed-by: Keyong Han <keyonghan@google.com>
Commit-Queue: Godofredo Contreras <godofredoc@google.com>