Remove any return ok from firebase test
This test was disabled because the invocation was flaky.
Since that time, we have started regularly running firebase tests
on LUCI and have not experienced the same flakes.

I fixed things up on the engine side yesterday, and have checked
that the last 10 runs are in fact passing on the engine dashboard.

Bug: https://github.com/flutter/flutter/issues/81216
Change-Id: I6ee9421a9c27db82636c3f72b691c2153ba38a90
Reviewed-on: https://flutter-review.googlesource.com/c/recipes/+/12900
Commit-Queue: Dan Field <dnfield@google.com>
Reviewed-by: Godofredo Contreras <godofredoc@google.com>